TAPSpriteBatch shouldn't be doing float math here.

+++ src/test/java/com/github/tommyettinger/batch/TextureArrayPolygonSpriteBatch.java
@@ -757,7 +757,7 @@ public class TextureArrayPolygonSpriteBatch extends PolygonSpriteBatch {
 
         //Calculate how many vertices and triangles will be put in the batch
         int triangleCount = (count / 20) * 6;
-        float verticesCount = ((float) (count / 5)) * 6;
+        int verticesCount = (count / 5) * 6;
         if (this.triangleIndex + triangleCount > triangles.length || this.vertexIndex + verticesCount > vertices.length)
             flush();
 
@@ -785,7 +785,7 @@ public class TextureArrayPolygonSpriteBatch extends PolygonSpriteBatch {
             vertices[vIn + 4] = spriteVertices[offsetIn + 4]; // v
             vertices[vIn + 5] = textureIndex; // texture index
         }
-        this.vertexIndex = (int) (vertexIndex + verticesCount);
+        this.vertexIndex = (vertexIndex + verticesCount);
     }
 
     @Override
